Do not serialise arguments to bricks
This saves a lot of trouble during initialisation, meaning that lots of
code can be removed. Bricks are not responsible for loading the rest of
the chain anymore: this is done by the frontend (once that is
implemented) and handlers are passed around as arguments. On the
flipside, this means that all bricks behind one frontend need to share a
virtual memory space. Not too much of a problem, since specialised
bricks can be used to overcome that limitation (like the TCP brick).

Also lowercase filenames of the minIni library.