Remove the 'connection thread' from TCP brick
It was created with the idea of being expanded into a prefetcher, but
that would be better as a totally separate prefetcher brick. Without
that, the connection thread only added code complexity and possible bugs
and race conditions. Hence the removal. Now the TCP brick shares
resources through a simple mutex (which is, execution time wise, not
worse than before).

Also make the TCP server use the new kfs_loadbrick module instead of
directly doing dynamic loading.