Pass state to every operation handler (no globals)
This is a very comprehensive patch because it requires a rewrite of a
significant part of the API and all bricks that implemented it.

With this patch, the initialisation method of a brick can return a void
* which will be passed along to every subsequent operation handler. FUSE
normally implements this by letting handlers calling get_context()
themselves, but for KennyFS this does not suffice because there can be
multiple different bricks (subvolumes) and thus multiple contexts.
Passing context explicitly allows handlers to be reentrant and makes the
code more straightforward.

Another complexity this difference between the KennyFS API and the FUSE
API introduces is that a translation layer is now needed to glue them
together: kfs_fuseoperglue.c.