Expect a handler for every operation, always
This saves the need for a if (oper->... == NULL) return -ENOSYS; in
every brick. Removes literally n * m runtime conditionals, where n is
the number of loaded bricks and m is the number of executed operations.
Provided, of course, that the compiler can not optimise this away (which
would be pretty impressive, given that bricks are dynamically loaded).

Of course, this does require every brick to provide a handler for every
operation, even if it makes no sense. In that case, the nosys_* handlers
(from kfs_nosys.c) can be used. These are simple return -ENOSYS
one-liners.