Add new KFS_DO_OPER macro to do KFS operations
It reduces the probability of a (otherwise easily introduced) bug:
forgetting to set the `priv' element of the context record to the
private data of the caller. This comes at the price of a very clunky
macro with lots of arguments that do not immediately make sense. A
matter of choosing explicitly confusing over implicitly confusing.