Verify that FUSE properly normalises the paths
This is one of the "FUSE invariants": all paths passed to handlers are
supposedly normalised and the first character is always a "/". Every
handler of the FUSE wrapper code now has an assert that checks for this
(and that apparently fails on heavy rm -rf, but that is another story).