Fix bug in kfs_sprintf: not enough space realloced
An early realloc caused less memory to be allocated than was expected,
which later on would cause writing out of array bounds. Also modified
the initial buffer to be a little more generous: 255 bytes are now
allocated upfront. This is the default maximum size for a (efficient)
smallbin in dlmalloc (most common malloc implementation).