Revert "Add consistency checks to POSIX brick"
This reverts commit 8b611544ac449d7599fc709809e0d9aef56e2d39.

Conflicts:

	posix_brick/kfs_brick_posix.c

Revert simulated by manually removing the consistency checks. Could not
be done automatically because the commit was polluted with some useful
new features.

Reason for reverting is that the tcp_server does not support this API
that requires it to keep track of all paths itself. Should be phased
out, anyway: it is a nice feature of the FUSE API for applications that
do not care about race conditions enough to associate paths with
filehandles internally, but that is not the case for KennyFS.